// src/interfaces/vehicles_types.ts
export default interface Vehicles {
        vehicles_data: {
          "name": string
          "model": string
          "vehicle_class": string
          "manufacturer": string
          "cost_in_credits": string
          "length": string
          "crew": string
          "passengers": string
          "max_atmosphering_speed": string
          "cargo_capacity": string
          "consumables": string
          "films": string[]
          "pilots": string[]
          "url": string
          "created": string
          "edited": string
        }
}